@@ -1,8 +1,9 @@
 package com.steelcolossus.mobiledev.mobileapplicationdevelopmentcoursework;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+import androidx.test.platform.app.InstrumentationRegistry;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,7 +22,7 @@ public class ExampleInstrumentedTest
     public void useAppContext()
     {
         // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getTargetContext();
+        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
         assertEquals("com.steelcolossus.mobiledev.mobileapplicationdevelopmentcoursework", appContext.getPackageName());
     }
