@@ -1,6 +1,6 @@
 package com.steelcolossus.mobiledev.mobileapplicationdevelopmentcoursework;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 
