@@ -1,8 +1,9 @@
 package com.steelcolossus.mobiledev.mobileapplicationdevelopmentcoursework;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.webkit.WebView;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class UserGuideActivity extends AppCompatActivity
 {
